refactor: extract NativeEventEmitter module helper
