+++ lib/commonjs/index.js
@@ -11,11 +11,9 @@ exports.removeListener = removeListener;
 exports.requestHint = requestHint;
 exports.startOtpListener = startOtpListener;
 exports.useOtpVerify = void 0;
-
 var _reactNative = require("react-native");
-
 var _react = require("react");
-
+var _nativeEventEmitterModule = require("./nativeEventEmitterModule");
 const LINKING_ERROR = `The package 'react-native-otp-verify' doesn't seem to be linked. Make sure: \n\n` + _reactNative.Platform.select({
   ios: "- You have run 'pod install'\n",
   default: ''
@@ -24,95 +22,65 @@ const RNOtpVerify = _reactNative.NativeModules.OtpVerify ? _reactNative.NativeMo
   get() {
     throw new Error(LINKING_ERROR);
   }
-
 });
-
-function getNativeEventEmitterModule() {
-  const nativeModule = _reactNative.NativeModules.OtpVerify !== null && _reactNative.NativeModules.OtpVerify !== void 0 ? _reactNative.NativeModules.OtpVerify : {};
-
-  if (typeof nativeModule.addListener !== 'function') {
-    nativeModule.addListener = () => {};
-  }
-
-  if (typeof nativeModule.removeListeners !== 'function') {
-    nativeModule.removeListeners = () => {};
-  }
-
-  return nativeModule;
-}
-
-const eventEmitter = new _reactNative.NativeEventEmitter(getNativeEventEmitterModule());
-
+const eventEmitter = new _reactNative.NativeEventEmitter((0, _nativeEventEmitterModule.getNativeEventEmitterModule)());
 async function getOtp() {
   if (_reactNative.Platform.OS === 'ios') {
     console.warn('Not Supported on iOS');
     return false;
   }
-
   return RNOtpVerify.getOtp();
 }
-
 function startOtpListener(handler) {
   return getOtp().then(() => addListener(handler));
 }
-
-const useOtpVerify = function () {
-  let {
-    numberOfDigits
-  } = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {
-    numberOfDigits: 0
-  };
+const useOtpVerify = ({
+  numberOfDigits
+} = {
+  numberOfDigits: 0
+}) => {
   const [message, setMessage] = (0, _react.useState)(null);
   const [otp, setOtp] = (0, _react.useState)(null);
   const [timeoutError, setTimeoutError] = (0, _react.useState)(false);
   const [hash, setHash] = (0, _react.useState)([]);
-
   const handleMessage = response => {
     if (response === 'Timeout Error.') {
       setTimeoutError(true);
     } else {
       setMessage(response);
-
       if (numberOfDigits && response) {
         const otpDigits = new RegExp(`(\\d{${numberOfDigits}})`, 'g').exec(response);
         if (otpDigits && otpDigits[1]) setOtp(otpDigits[1]);
       }
     }
   };
-
   (0, _react.useEffect)(() => {
     if (_reactNative.Platform.OS === 'ios') {
       console.warn('Not Supported on iOS');
       return;
     }
-
     getHash().then(setHash);
     startOtpListener(handleMessage);
     return () => {
       removeListener();
     };
   }, []);
-
   const startListener = () => {
     if (_reactNative.Platform.OS === 'ios') {
       console.warn('Not Supported on iOS');
       return;
     }
-
     setOtp('');
     setMessage('');
     startOtpListener(handleMessage);
   };
-
   const stopListener = () => {
     if (_reactNative.Platform.OS === 'ios') {
       console.warn('Not Supported on iOS');
       return;
     }
-
     removeListener();
   };
-
   return {
     otp,
     message,
@@ -122,35 +90,27 @@ const useOtpVerify = function () {
     startListener
   };
 };
-
 exports.useOtpVerify = useOtpVerify;
-
 async function getHash() {
   if (_reactNative.Platform.OS === 'ios') {
     console.warn('Not Supported on iOS');
     return [];
   }
-
   return RNOtpVerify.getHash();
 }
-
 async function requestHint() {
   if (_reactNative.Platform.OS === 'ios') {
     console.warn('Not Supported on iOS');
     return '';
   }
-
   return RNOtpVerify.requestHint();
 }
-
 function addListener(handler) {
   return eventEmitter.addListener('com.faizalshap.otpVerify:otpReceived', handler);
 }
-
 function removeListener() {
   return eventEmitter.removeAllListeners('com.faizalshap.otpVerify:otpReceived');
 }
-
 const OtpVerify = {
   getOtp,
   getHash,
@@ -159,6 +119,5 @@ const OtpVerify = {
   startOtpListener,
   requestHint
 };
-var _default = OtpVerify;
-exports.default = _default;
+var _default = exports.default = OtpVerify;
 //# sourceMappingURL=index.js.map
