+++ lib/module/index.js
@@ -1,5 +1,6 @@
 import { NativeEventEmitter, NativeModules, Platform } from 'react-native';
 import { useEffect, useState } from 'react';
+import { getNativeEventEmitterModule } from './nativeEventEmitterModule';
 const LINKING_ERROR = `The package 'react-native-otp-verify' doesn't seem to be linked. Make sure: \n\n` + Platform.select({
   ios: "- You have run 'pod install'\n",
   default: ''
@@ -8,92 +9,65 @@ const RNOtpVerify = NativeModules.OtpVerify ? NativeModules.OtpVerify : new Prox
   get() {
     throw new Error(LINKING_ERROR);
   }
-
 });
-
-function getNativeEventEmitterModule() {
-  const nativeModule = NativeModules.OtpVerify ?? {};
-
-  if (typeof nativeModule.addListener !== 'function') {
-    nativeModule.addListener = () => {};
-  }
-
-  if (typeof nativeModule.removeListeners !== 'function') {
-    nativeModule.removeListeners = () => {};
-  }
-
-  return nativeModule;
-}
-
 const eventEmitter = new NativeEventEmitter(getNativeEventEmitterModule());
 export async function getOtp() {
   if (Platform.OS === 'ios') {
     console.warn('Not Supported on iOS');
     return false;
   }
-
   return RNOtpVerify.getOtp();
 }
 export function startOtpListener(handler) {
   return getOtp().then(() => addListener(handler));
 }
-export const useOtpVerify = function () {
-  let {
-    numberOfDigits
-  } = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {
-    numberOfDigits: 0
-  };
+export const useOtpVerify = ({
+  numberOfDigits
+} = {
+  numberOfDigits: 0
+}) => {
   const [message, setMessage] = useState(null);
   const [otp, setOtp] = useState(null);
   const [timeoutError, setTimeoutError] = useState(false);
   const [hash, setHash] = useState([]);
-
   const handleMessage = response => {
     if (response === 'Timeout Error.') {
       setTimeoutError(true);
     } else {
       setMessage(response);
-
       if (numberOfDigits && response) {
         const otpDigits = new RegExp(`(\\d{${numberOfDigits}})`, 'g').exec(response);
         if (otpDigits && otpDigits[1]) setOtp(otpDigits[1]);
       }
     }
   };
-
   useEffect(() => {
     if (Platform.OS === 'ios') {
       console.warn('Not Supported on iOS');
       return;
     }
-
     getHash().then(setHash);
     startOtpListener(handleMessage);
     return () => {
       removeListener();
     };
   }, []);
-
   const startListener = () => {
     if (Platform.OS === 'ios') {
       console.warn('Not Supported on iOS');
       return;
     }
-
     setOtp('');
     setMessage('');
     startOtpListener(handleMessage);
   };
-
   const stopListener = () => {
     if (Platform.OS === 'ios') {
       console.warn('Not Supported on iOS');
       return;
     }
-
     removeListener();
   };
-
   return {
     otp,
     message,
@@ -108,7 +82,6 @@ export async function getHash() {
     console.warn('Not Supported on iOS');
     return [];
   }
-
   return RNOtpVerify.getHash();
 }
 export async function requestHint() {
@@ -116,7 +89,6 @@ export async function requestHint() {
     console.warn('Not Supported on iOS');
     return '';
   }
-
   return RNOtpVerify.requestHint();
 }
 export function addListener(handler) {
